refactor(client): define custom `useAuth{State,}` hooks
