+++ .biome.jsonc
@@ -54,7 +54,7 @@
     "linter": { "rules": { "style": { "useImportType": "off" } } },
   }, {
     "include": [
-      "apps/client/src/app/page.tsx",
+      "apps/client/src/hooks/useAuth.ts",
     ],
     "linter": { "rules": { "correctness": { "useExhaustiveDependencies": "off" } } },
   }, {
