+++ apps/client/src/state/auth/reducer.ts
@@ -3,8 +3,6 @@ import type { AuthAction, AuthState } from 'client/state/auth/types'
 export const initialAuthState: AuthState = {
   email: '',
   emailSent: false,
-  authenticated: false,
-  signedMessage: undefined,
 }
 
 export const authReducer = (state: AuthState, action: AuthAction): AuthState => {
@@ -13,10 +11,6 @@ export const authReducer = (state: AuthState, action: AuthAction): AuthState => 
       return { ...state, email: action.payload }
     case 'EMAIL_SENT':
       return { ...state, emailSent: true }
-    case 'AUTHENTICATED':
-      return { ...state, authenticated: true }
-    case 'SIGNED_MESSAGE':
-      return { ...state, signedMessage: action.payload }
     case 'LOGOUT':
       return initialAuthState
     default:
